Decouples ROS from the core LOAM algorithms. - Maintains original API. - Core LOAM algorithm now only depend on PCL and Eigen. - Uses std::chrono instead of ROS Time (bit-exact change).
