+++ .gitignore
@@ -9,3 +9,4 @@ CMakeFiles
 cmake_install.cmake
 Makefile
 .idea
+.vs
