+++ include/loam_velodyne/BasicTransformMaintenance.h
@@ -1,3 +1,4 @@
+#pragma once
 // Copyright 2013, Ji Zhang, Carnegie Mellon University
 // Further contributions copyright (c) 2016, Southwest Research Institute
 // All rights reserved.
@@ -30,51 +31,36 @@
 //   J. Zhang and S. Singh. LOAM: Lidar Odometry and Mapping in Real-time.
 //     Robotics: Science and Systems Conference (RSS). Berkeley, CA, July 2014.
 
-#ifndef LOAM_COMMON_H
-#define LOAM_COMMON_H
+#include "Twist.h"
 
-#include <ros/ros.h>
-#include <sensor_msgs/PointCloud2.h>
-#include <pcl_conversions/pcl_conversions.h>
-#include <pcl/point_types.h>
-#include "time_utils.h"
+namespace loam
+{
 
-namespace loam {
-
-/** \brief Construct a new point cloud message from the specified information and publish it via the given publisher.
+/** \brief Implementation of the LOAM transformation maintenance component.
  *
- * @tparam PointT the point type
- * @param publisher the publisher instance
- * @param cloud the cloud to publish
- * @param stamp the time stamp of the cloud message
- * @param frameID the message frame ID
  */
-template <typename PointT>
-inline void publishCloudMsg(ros::Publisher& publisher,
-                            const pcl::PointCloud<PointT>& cloud,
-                            const ros::Time& stamp,
-                            std::string frameID) {
-  sensor_msgs::PointCloud2 msg;
-  pcl::toROSMsg(cloud, msg);
-  msg.header.stamp = stamp;
-  msg.header.frame_id = frameID;
-  publisher.publish(msg);
-}
-
-
-// ROS time adapters
-inline Time fromROSTime(ros::Time const& rosTime)
+class BasicTransformMaintenance
 {
-  auto epoch = std::chrono::system_clock::time_point();
-  auto since_epoch = std::chrono::seconds(rosTime.sec) + std::chrono::nanoseconds(rosTime.nsec);
-  return epoch + since_epoch;
-}
+public:
+   void updateOdometry(double pitch, double yaw, double roll, double x, double y, double z);
+   void updateMappingTransform(Twist const& transformAftMapped, Twist const& transformBefMapped);
+   void updateMappingTransform(double pitch, double yaw, double roll,
+      double x, double y, double z,
+      double twist_rot_x, double twist_rot_y, double twist_rot_z,
+      double twist_pos_x, double twist_pos_y, double twist_pos_z);
 
-inline ros::Time toROSTime(Time const& time_point)
-{
-  return ros::Time().fromNSec(std::chrono::duration_cast<std::chrono::nanoseconds>(time_point.time_since_epoch()).count());
-}
+   void transformAssociateToMap();
+
+   // result accessor
+   auto const& transformMapped() const { return _transformMapped; }
+
+private:
+   float _transformSum[6]{};
+   float _transformIncre[6]{};
+   float _transformMapped[6]{};
+   float _transformBefMapped[6]{};
+   float _transformAftMapped[6]{};
+};
 
 } // end namespace loam
 
-#endif // LOAM_COMMON_H
