+++ src/lib/CMakeLists.txt
@@ -1,8 +1,12 @@
 add_library(loam
             math_utils.h
             ScanRegistration.cpp
+            BasicScanRegistration.cpp
             MultiScanRegistration.cpp
             LaserOdometry.cpp
+            BasicLaserOdometry.cpp
             LaserMapping.cpp
-            TransformMaintenance.cpp)
+            BasicLaserMapping.cpp
+            TransformMaintenance.cpp
+            BasicTransformMaintenance.cpp)
 target_link_libraries(loam ${catkin_LIBRARIES} ${PCL_LIBRARIES})
