+++ src/lib/ScanRegistration.cpp
@@ -33,37 +33,15 @@
 #include "loam_velodyne/ScanRegistration.h"
 #include "math_utils.h"
 
-#include <pcl/filters/voxel_grid.h>
 #include <tf/transform_datatypes.h>
 
 
 namespace loam {
 
 
-RegistrationParams::RegistrationParams(const float& scanPeriod_,
-                                       const int& imuHistorySize_,
-                                       const int& nFeatureRegions_,
-                                       const int& curvatureRegion_,
-                                       const int& maxCornerSharp_,
-                                       const int& maxSurfaceFlat_,
-                                       const float& lessFlatFilterSize_,
-                                       const float& surfaceCurvatureThreshold_)
-    : scanPeriod(scanPeriod_),
-      imuHistorySize(imuHistorySize_),
-      nFeatureRegions(nFeatureRegions_),
-      curvatureRegion(curvatureRegion_),
-      maxCornerSharp(maxCornerSharp_),
-      maxCornerLessSharp(10 * maxCornerSharp_),
-      maxSurfaceFlat(maxSurfaceFlat_),
-      lessFlatFilterSize(lessFlatFilterSize_),
-      surfaceCurvatureThreshold(surfaceCurvatureThreshold_)
+
+bool ScanRegistration::parseParams(const ros::NodeHandle& nh, RegistrationParams& config_out) 
 {
-
-};
-
-
-
-bool RegistrationParams::parseParams(const ros::NodeHandle& nh) {
   bool success = true;
   int iParam = 0;
   float fParam = 0;
@@ -73,7 +51,7 @@ bool RegistrationParams::parseParams(const ros::NodeHandle& nh) {
       ROS_ERROR("Invalid scanPeriod parameter: %f (expected > 0)", fParam);
       success = false;
     } else {
-      scanPeriod = fParam;
+        config_out.scanPeriod = fParam;
       ROS_INFO("Set scanPeriod: %g", fParam);
     }
   }
@@ -83,7 +61,7 @@ bool RegistrationParams::parseParams(const ros::NodeHandle& nh) {
       ROS_ERROR("Invalid imuHistorySize parameter: %d (expected >= 1)", iParam);
       success = false;
     } else {
-      imuHistorySize = iParam;
+        config_out.imuHistorySize = iParam;
       ROS_INFO("Set imuHistorySize: %d", iParam);
     }
   }
@@ -93,7 +71,7 @@ bool RegistrationParams::parseParams(const ros::NodeHandle& nh) {
       ROS_ERROR("Invalid featureRegions parameter: %d (expected >= 1)", iParam);
       success = false;
     } else {
-      nFeatureRegions = iParam;
+        config_out.nFeatureRegions = iParam;
       ROS_INFO("Set nFeatureRegions: %d", iParam);
     }
   }
@@ -103,7 +81,7 @@ bool RegistrationParams::parseParams(const ros::NodeHandle& nh) {
       ROS_ERROR("Invalid curvatureRegion parameter: %d (expected >= 1)", iParam);
       success = false;
     } else {
-      curvatureRegion = iParam;
+        config_out.curvatureRegion = iParam;
       ROS_INFO("Set curvatureRegion: +/- %d", iParam);
     }
   }
@@ -113,18 +91,18 @@ bool RegistrationParams::parseParams(const ros::NodeHandle& nh) {
       ROS_ERROR("Invalid maxCornerSharp parameter: %d (expected >= 1)", iParam);
       success = false;
     } else {
-      maxCornerSharp = iParam;
-      maxCornerLessSharp = 10 * iParam;
-      ROS_INFO("Set maxCornerSharp / less sharp: %d / %d", iParam, maxCornerLessSharp);
+        config_out.maxCornerSharp = iParam;
+        config_out.maxCornerLessSharp = 10 * iParam;
+      ROS_INFO("Set maxCornerSharp / less sharp: %d / %d", iParam, config_out.maxCornerLessSharp);
     }
   }
 
   if (nh.getParam("maxCornerLessSharp", iParam)) {
-    if (iParam < maxCornerSharp) {
-      ROS_ERROR("Invalid maxCornerLessSharp parameter: %d (expected >= %d)", iParam, maxCornerSharp);
+    if (iParam < config_out.maxCornerSharp) {
+      ROS_ERROR("Invalid maxCornerLessSharp parameter: %d (expected >= %d)", iParam, config_out.maxCornerSharp);
       success = false;
     } else {
-      maxCornerLessSharp = iParam;
+        config_out.maxCornerLessSharp = iParam;
       ROS_INFO("Set maxCornerLessSharp: %d", iParam);
     }
   }
@@ -134,7 +112,7 @@ bool RegistrationParams::parseParams(const ros::NodeHandle& nh) {
       ROS_ERROR("Invalid maxSurfaceFlat parameter: %d (expected >= 1)", iParam);
       success = false;
     } else {
-      maxSurfaceFlat = iParam;
+        config_out.maxSurfaceFlat = iParam;
       ROS_INFO("Set maxSurfaceFlat: %d", iParam);
     }
   }
@@ -144,7 +122,7 @@ bool RegistrationParams::parseParams(const ros::NodeHandle& nh) {
       ROS_ERROR("Invalid surfaceCurvatureThreshold parameter: %f (expected >= 0.001)", fParam);
       success = false;
     } else {
-      surfaceCurvatureThreshold = fParam;
+        config_out.surfaceCurvatureThreshold = fParam;
       ROS_INFO("Set surfaceCurvatureThreshold: %g", fParam);
     }
   }
@@ -154,7 +132,7 @@ bool RegistrationParams::parseParams(const ros::NodeHandle& nh) {
       ROS_ERROR("Invalid lessFlatFilterSize parameter: %f (expected >= 0.001)", fParam);
       success = false;
     } else {
-      lessFlatFilterSize = fParam;
+        config_out.lessFlatFilterSize = fParam;
       ROS_INFO("Set lessFlatFilterSize: %g", fParam);
     }
   }
@@ -162,54 +140,21 @@ bool RegistrationParams::parseParams(const ros::NodeHandle& nh) {
   return success;
 }
 
-
-
-
-
-
-ScanRegistration::ScanRegistration(const RegistrationParams& config)
-      : _config(config),
-        _sweepStart(),
-        _scanTime(),
-        _imuStart(),
-        _imuCur(),
-        _imuIdx(0),
-        _imuHistory(_config.imuHistorySize),
-        _laserCloud(),
-        _cornerPointsSharp(),
-        _cornerPointsLessSharp(),
-        _surfacePointsFlat(),
-        _surfacePointsLessFlat(),
-        _imuTrans(4, 1),
-        _regionCurvature(),
-        _regionLabel(),
-        _regionSortIndices(),
-        _scanNeighborPicked()
+bool ScanRegistration::setupROS(ros::NodeHandle& node, ros::NodeHandle& privateNode, RegistrationParams& config_out)
 {
-
-}
-
-
-
-bool ScanRegistration::setup(ros::NodeHandle& node,
-                             ros::NodeHandle& privateNode)
-{
-  if (!_config.parseParams(privateNode)) {
+  if (!parseParams(privateNode, config_out))
     return false;
-  }
-  _imuHistory.ensureCapacity(_config.imuHistorySize);
 
   // subscribe to IMU topic
   _subImu = node.subscribe<sensor_msgs::Imu>("/imu/data", 50, &ScanRegistration::handleIMUMessage, this);
 
-
   // advertise scan registration topics
-  _pubLaserCloud = node.advertise<sensor_msgs::PointCloud2>("/velodyne_cloud_2", 2);
-  _pubCornerPointsSharp = node.advertise<sensor_msgs::PointCloud2>("/laser_cloud_sharp", 2);
+  _pubLaserCloud            = node.advertise<sensor_msgs::PointCloud2>("/velodyne_cloud_2", 2);
+  _pubCornerPointsSharp     = node.advertise<sensor_msgs::PointCloud2>("/laser_cloud_sharp", 2);
   _pubCornerPointsLessSharp = node.advertise<sensor_msgs::PointCloud2>("/laser_cloud_less_sharp", 2);
-  _pubSurfPointsFlat = node.advertise<sensor_msgs::PointCloud2>("/laser_cloud_flat", 2);
-  _pubSurfPointsLessFlat = node.advertise<sensor_msgs::PointCloud2>("/laser_cloud_less_flat", 2);
-  _pubImuTrans = node.advertise<sensor_msgs::PointCloud2>("/imu_trans", 5);
+  _pubSurfPointsFlat        = node.advertise<sensor_msgs::PointCloud2>("/laser_cloud_flat", 2);
+  _pubSurfPointsLessFlat    = node.advertise<sensor_msgs::PointCloud2>("/laser_cloud_less_flat", 2);
+  _pubImuTrans              = node.advertise<sensor_msgs::PointCloud2>("/imu_trans", 5);
 
   return true;
 }
@@ -218,362 +163,39 @@ bool ScanRegistration::setup(ros::NodeHandle& node,
 
 void ScanRegistration::handleIMUMessage(const sensor_msgs::Imu::ConstPtr& imuIn)
 {
-  double roll, pitch, yaw;
   tf::Quaternion orientation;
   tf::quaternionMsgToTF(imuIn->orientation, orientation);
+  double roll, pitch, yaw;
   tf::Matrix3x3(orientation).getRPY(roll, pitch, yaw);
 
   Vector3 acc;
   acc.x() = float(imuIn->linear_acceleration.y - sin(roll) * cos(pitch) * 9.81);
   acc.y() = float(imuIn->linear_acceleration.z - cos(roll) * cos(pitch) * 9.81);
-  acc.z() = float(imuIn->linear_acceleration.x + sin(pitch) * 9.81);
+  acc.z() = float(imuIn->linear_acceleration.x + sin(pitch)             * 9.81);
 
   IMUState newState;
-  newState.stamp = imuIn->header.stamp;
+  newState.stamp = fromROSTime( imuIn->header.stamp);
   newState.roll = roll;
   newState.pitch = pitch;
   newState.yaw = yaw;
   newState.acceleration = acc;
 
-  if (_imuHistory.size() > 0) {
-    // accumulate IMU position and velocity over time
-    rotateZXY(acc, newState.roll, newState.pitch, newState.yaw);
-
-    const IMUState& prevState = _imuHistory.last();
-    float timeDiff = float((newState.stamp - prevState.stamp).toSec());
-    newState.position = prevState.position
-                        + (prevState.velocity * timeDiff)
-                        + (0.5 * acc * timeDiff * timeDiff);
-    newState.velocity = prevState.velocity
-                        + acc * timeDiff;
-  }
-
-  _imuHistory.push(newState);
+  updateIMUData(acc, newState);
 }
-
-
-
-void ScanRegistration::reset(const ros::Time& scanTime,
-                             const bool& newSweep)
-{
-  _scanTime = scanTime;
-
-  // re-initialize IMU start index and state
-  _imuIdx = 0;
-  if (hasIMUData()) {
-    interpolateIMUStateFor(0, _imuStart);
-  }
-
-  // clear internal cloud buffers at the beginning of a sweep
-  if (newSweep) {
-    _sweepStart = scanTime;
-
-    // clear cloud buffers
-    _laserCloud.clear();
-    _cornerPointsSharp.clear();
-    _cornerPointsLessSharp.clear();
-    _surfacePointsFlat.clear();
-    _surfacePointsLessFlat.clear();
-
-    // clear scan indices vector
-    _scanIndices.clear();
-  }
-}
-
-
-
-void ScanRegistration::setIMUTransformFor(const float& relTime)
-{
-  interpolateIMUStateFor(relTime, _imuCur);
-
-  float relSweepTime = (_scanTime - _sweepStart).toSec() + relTime;
-  _imuPositionShift = _imuCur.position - _imuStart.position - _imuStart.velocity * relSweepTime;
-}
-
-
-
-void ScanRegistration::transformToStartIMU(pcl::PointXYZI& point)
-{
-  // rotate point to global IMU system
-  rotateZXY(point, _imuCur.roll, _imuCur.pitch, _imuCur.yaw);
-
-  // add global IMU position shift
-  point.x += _imuPositionShift.x();
-  point.y += _imuPositionShift.y();
-  point.z += _imuPositionShift.z();
-
-  // rotate point back to local IMU system relative to the start IMU state
-  rotateYXZ(point, -_imuStart.yaw, -_imuStart.pitch, -_imuStart.roll);
-}
-
-
-
-void ScanRegistration::interpolateIMUStateFor(const float &relTime,
-                                              IMUState &outputState)
-{
-  double timeDiff = (_scanTime - _imuHistory[_imuIdx].stamp).toSec() + relTime;
-  while (_imuIdx < _imuHistory.size() - 1 && timeDiff > 0) {
-    _imuIdx++;
-    timeDiff = (_scanTime - _imuHistory[_imuIdx].stamp).toSec() + relTime;
-  }
-
-  if (_imuIdx == 0 || timeDiff > 0) {
-    outputState = _imuHistory[_imuIdx];
-  } else {
-    float ratio = -timeDiff / (_imuHistory[_imuIdx].stamp - _imuHistory[_imuIdx - 1].stamp).toSec();
-    IMUState::interpolate(_imuHistory[_imuIdx], _imuHistory[_imuIdx - 1], ratio, outputState);
-  }
-}
-
-
-
-void ScanRegistration::extractFeatures(const uint16_t& beginIdx)
-{
-  // extract features from individual scans
-  size_t nScans = _scanIndices.size();
-  for (size_t i = beginIdx; i < nScans; i++) {
-    pcl::PointCloud<pcl::PointXYZI>::Ptr surfPointsLessFlatScan(new pcl::PointCloud<pcl::PointXYZI>);
-    size_t scanStartIdx = _scanIndices[i].first;
-    size_t scanEndIdx = _scanIndices[i].second;
-
-    // skip empty scans
-    if (scanEndIdx <= scanStartIdx + 2 * _config.curvatureRegion) {
-      continue;
-    }
-
-    // Quick&Dirty fix for relative point time calculation without IMU data
-    /*float scanSize = scanEndIdx - scanStartIdx + 1;
-    for (int j = scanStartIdx; j <= scanEndIdx; j++) {
-      _laserCloud[j].intensity = i + _scanPeriod * (j - scanStartIdx) / scanSize;
-    }*/
-
-    // reset scan buffers
-    setScanBuffersFor(scanStartIdx, scanEndIdx);
-
-    // extract features from equally sized scan regions
-    for (int j = 0; j < _config.nFeatureRegions; j++) {
-      size_t sp = ((scanStartIdx + _config.curvatureRegion) * (_config.nFeatureRegions - j)
-                   + (scanEndIdx - _config.curvatureRegion) * j) / _config.nFeatureRegions;
-      size_t ep = ((scanStartIdx + _config.curvatureRegion) * (_config.nFeatureRegions - 1 - j)
-                   + (scanEndIdx - _config.curvatureRegion) * (j + 1)) / _config.nFeatureRegions - 1;
-
-      // skip empty regions
-      if (ep <= sp) {
-        continue;
-      }
-
-      size_t regionSize = ep - sp + 1;
-
-      // reset region buffers
-      setRegionBuffersFor(sp, ep);
-
-
-      // extract corner features
-      int largestPickedNum = 0;
-      for (size_t k = regionSize; k > 0 && largestPickedNum < _config.maxCornerLessSharp;) {
-        size_t idx = _regionSortIndices[--k];
-        size_t scanIdx = idx - scanStartIdx;
-        size_t regionIdx = idx - sp;
-
-        if (_scanNeighborPicked[scanIdx] == 0 &&
-            _regionCurvature[regionIdx] > _config.surfaceCurvatureThreshold) {
-
-          largestPickedNum++;
-          if (largestPickedNum <= _config.maxCornerSharp) {
-            _regionLabel[regionIdx] = CORNER_SHARP;
-            _cornerPointsSharp.push_back(_laserCloud[idx]);
-          } else {
-            _regionLabel[regionIdx] = CORNER_LESS_SHARP;
-          }
-          _cornerPointsLessSharp.push_back(_laserCloud[idx]);
-
-          markAsPicked(idx, scanIdx);
-        }
-      }
-
-      // extract flat surface features
-      int smallestPickedNum = 0;
-      for (int k = 0; k < regionSize && smallestPickedNum < _config.maxSurfaceFlat; k++) {
-        size_t idx = _regionSortIndices[k];
-        size_t scanIdx = idx - scanStartIdx;
-        size_t regionIdx = idx - sp;
-
-        if (_scanNeighborPicked[scanIdx] == 0 &&
-            _regionCurvature[regionIdx] < _config.surfaceCurvatureThreshold) {
-
-          smallestPickedNum++;
-          _regionLabel[regionIdx] = SURFACE_FLAT;
-          _surfacePointsFlat.push_back(_laserCloud[idx]);
-
-          markAsPicked(idx, scanIdx);
-        }
-      }
-
-      // extract less flat surface features
-      for (int k = 0; k < regionSize; k++) {
-        if (_regionLabel[k] <= SURFACE_LESS_FLAT) {
-          surfPointsLessFlatScan->push_back(_laserCloud[sp + k]);
-        }
-      }
-    }
-
-    // down size less flat surface point cloud of current scan
-    pcl::PointCloud<pcl::PointXYZI> surfPointsLessFlatScanDS;
-    pcl::VoxelGrid<pcl::PointXYZI> downSizeFilter;
-    downSizeFilter.setInputCloud(surfPointsLessFlatScan);
-    downSizeFilter.setLeafSize(_config.lessFlatFilterSize, _config.lessFlatFilterSize, _config.lessFlatFilterSize);
-    downSizeFilter.filter(surfPointsLessFlatScanDS);
-
-    _surfacePointsLessFlat += surfPointsLessFlatScanDS;
-  }
-}
-
-
-
-void ScanRegistration::setRegionBuffersFor(const size_t& startIdx,
-                                           const size_t& endIdx)
-{
-  // resize buffers
-  size_t regionSize = endIdx - startIdx + 1;
-  _regionCurvature.resize(regionSize);
-  _regionSortIndices.resize(regionSize);
-  _regionLabel.assign(regionSize, SURFACE_LESS_FLAT);
-
-  // calculate point curvatures and reset sort indices
-  float pointWeight = -2 * _config.curvatureRegion;
-
-  for (size_t i = startIdx, regionIdx = 0; i <= endIdx; i++, regionIdx++) {
-    float diffX = pointWeight * _laserCloud[i].x;
-    float diffY = pointWeight * _laserCloud[i].y;
-    float diffZ = pointWeight * _laserCloud[i].z;
-
-    for (int j = 1; j <= _config.curvatureRegion; j++) {
-      diffX += _laserCloud[i + j].x + _laserCloud[i - j].x;
-      diffY += _laserCloud[i + j].y + _laserCloud[i - j].y;
-      diffZ += _laserCloud[i + j].z + _laserCloud[i - j].z;
-    }
-
-    _regionCurvature[regionIdx] = diffX * diffX + diffY * diffY + diffZ * diffZ;
-    _regionSortIndices[regionIdx] = i;
-  }
-
-  // sort point curvatures
-  for (size_t i = 1; i < regionSize; i++) {
-    for (size_t j = i; j >= 1; j--) {
-      if (_regionCurvature[_regionSortIndices[j] - startIdx] < _regionCurvature[_regionSortIndices[j - 1] - startIdx]) {
-        std::swap(_regionSortIndices[j], _regionSortIndices[j - 1]);
-      }
-    }
-  }
-}
-
-
-
-void ScanRegistration::setScanBuffersFor(const size_t& startIdx,
-                                         const size_t& endIdx)
-{
-  // resize buffers
-  size_t scanSize = endIdx - startIdx + 1;
-  _scanNeighborPicked.assign(scanSize, 0);
-
-  // mark unreliable points as picked
-  for (size_t i = startIdx + _config.curvatureRegion; i < endIdx - _config.curvatureRegion; i++) {
-    const pcl::PointXYZI& previousPoint = (_laserCloud[i - 1]);
-    const pcl::PointXYZI& point = (_laserCloud[i]);
-    const pcl::PointXYZI& nextPoint = (_laserCloud[i + 1]);
-
-    float diffNext = calcSquaredDiff(nextPoint, point);
-
-    if (diffNext > 0.1) {
-      float depth1 = calcPointDistance(point);
-      float depth2 = calcPointDistance(nextPoint);
-
-      if (depth1 > depth2) {
-        float weighted_distance = std::sqrt(calcSquaredDiff(nextPoint, point, depth2 / depth1)) / depth2;
-
-        if (weighted_distance < 0.1) {
-          std::fill_n(&_scanNeighborPicked[i - startIdx - _config.curvatureRegion], _config.curvatureRegion + 1, 1);
-
-          continue;
-        }
-      } else {
-        float weighted_distance = std::sqrt(calcSquaredDiff(point, nextPoint, depth1 / depth2)) / depth1;
-
-        if (weighted_distance < 0.1) {
-          std::fill_n(&_scanNeighborPicked[i - startIdx + 1], _config.curvatureRegion + 1, 1);
-        }
-      }
-    }
-
-    float diffPrevious = calcSquaredDiff(point, previousPoint);
-    float dis = calcSquaredPointDistance(point);
-
-    if (diffNext > 0.0002 * dis && diffPrevious > 0.0002 * dis) {
-      _scanNeighborPicked[i - startIdx] = 1;
-    }
-  }
-}
-
-
-
-void ScanRegistration::markAsPicked(const size_t& cloudIdx,
-                                    const size_t& scanIdx)
-{
-  _scanNeighborPicked[scanIdx] = 1;
-
-  for (int i = 1; i <= _config.curvatureRegion; i++) {
-    if (calcSquaredDiff(_laserCloud[cloudIdx + i], _laserCloud[cloudIdx + i - 1]) > 0.05) {
-      break;
-    }
-
-    _scanNeighborPicked[scanIdx + i] = 1;
-  }
-
-  for (int i = 1; i <= _config.curvatureRegion; i++) {
-    if (calcSquaredDiff(_laserCloud[cloudIdx - i], _laserCloud[cloudIdx - i + 1]) > 0.05) {
-      break;
-    }
-
-    _scanNeighborPicked[scanIdx - i] = 1;
-  }
-}
-
 
 
 void ScanRegistration::publishResult()
 {
+  auto sweepStartTime = toROSTime(sweepStart());
   // publish full resolution and feature point clouds
-  publishCloudMsg(_pubLaserCloud,            _laserCloud,            _sweepStart, "/camera");
-  publishCloudMsg(_pubCornerPointsSharp,     _cornerPointsSharp,     _sweepStart, "/camera");
-  publishCloudMsg(_pubCornerPointsLessSharp, _cornerPointsLessSharp, _sweepStart, "/camera");
-  publishCloudMsg(_pubSurfPointsFlat,        _surfacePointsFlat,     _sweepStart, "/camera");
-  publishCloudMsg(_pubSurfPointsLessFlat,    _surfacePointsLessFlat, _sweepStart, "/camera");
-
+  publishCloudMsg(_pubLaserCloud, laserCloud(), sweepStartTime, "/camera");
+  publishCloudMsg(_pubCornerPointsSharp, cornerPointsSharp(), sweepStartTime, "/camera");
+  publishCloudMsg(_pubCornerPointsLessSharp, cornerPointsLessSharp(), sweepStartTime, "/camera");
+  publishCloudMsg(_pubSurfPointsFlat, surfacePointsFlat(), sweepStartTime, "/camera");
+  publishCloudMsg(_pubSurfPointsLessFlat, surfacePointsLessFlat(), sweepStartTime, "/camera");
 
   // publish corresponding IMU transformation information
-  _imuTrans[0].x = _imuStart.pitch.rad();
-  _imuTrans[0].y = _imuStart.yaw.rad();
-  _imuTrans[0].z = _imuStart.roll.rad();
-
-  _imuTrans[1].x = _imuCur.pitch.rad();
-  _imuTrans[1].y = _imuCur.yaw.rad();
-  _imuTrans[1].z = _imuCur.roll.rad();
-
-  Vector3 imuShiftFromStart = _imuPositionShift;
-  rotateYXZ(imuShiftFromStart, -_imuStart.yaw, -_imuStart.pitch, -_imuStart.roll);
-
-  _imuTrans[2].x = imuShiftFromStart.x();
-  _imuTrans[2].y = imuShiftFromStart.y();
-  _imuTrans[2].z = imuShiftFromStart.z();
-
-  Vector3 imuVelocityFromStart = _imuCur.velocity - _imuStart.velocity;
-  rotateYXZ(imuVelocityFromStart, -_imuStart.yaw, -_imuStart.pitch, -_imuStart.roll);
-
-  _imuTrans[3].x = imuVelocityFromStart.x();
-  _imuTrans[3].y = imuVelocityFromStart.y();
-  _imuTrans[3].z = imuVelocityFromStart.z();
-
-  publishCloudMsg(_pubImuTrans, _imuTrans, _sweepStart, "/camera");
+  publishCloudMsg(_pubImuTrans, imuTransform(), sweepStartTime, "/camera");
 }
 
 } // end namespace loam
